pyproject.toml
[project]
name = "rav"
version = "0.0.9"
authors = [
  { name="Justin Mitchel", email="hello@teamcfe.com" },
]
description = "A cross-platform Python CLI to shortcut tp command-line commands. Inspired by Makefiles and npm scripts."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ['fire', 'PyYAML', 'rich']

[project.urls]
"Homepage" = "https://github.com/jmitchel3/rev"
"Bug Tracker" = "https://github.com/jmitchel3/rev/issues"